+++ packages/likec4/src/cli/export/utils/takeScreenshot.ts
@@ -1,71 +1,82 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import type { DiagramView } from '@likec4/core'
+/// <reference lib="DOM" />
+import type { DiagramView, NonEmptyArray } from '@likec4/core'
 import { resolve } from 'node:path'
 import k from 'picocolors'
-import type { Browser, Page } from 'playwright-core'
+import type { BrowserContext, Page } from 'playwright-core'
+import { isString } from 'remeda'
 import type { Logger } from 'vite'
 
 type TakeScreenshotParams = {
-  page: Page
-  pageUrl: (view: DiagramView) => string
-  outputDir: string
+  browserContext: BrowserContext
+  views: NonEmptyArray<DiagramView>
+  output: string
   logger: Logger
 }
 
-export function mkTakeScreenshotFn({ page, pageUrl, outputDir, logger }: TakeScreenshotParams) {
-  return async function takeScreenshot(view: DiagramView) {
-    const padding = 24
-    const url = pageUrl(view)
-    logger.info(`${k.cyan('export')} ${view.id} ${k.underline(k.dim(url))}`)
+export async function takeScreenshot({ browserContext, views, output, logger }: TakeScreenshotParams) {
+  const padding = 20
 
+  let page: Page | undefined
+
+  const succeed = [] as Array<{ view: DiagramView; path: string }>
+
+  for (const view of views) {
     try {
+      const url = `/export/${encodeURIComponent(view.id)}`
+      logger.info(k.cyan(url))
+
+      page ??= await browserContext.newPage()
+
       await page.setViewportSize({
         width: view.width + padding * 2 + 4,
         height: view.height + padding * 2 + 4
       })
-    } catch (error: unknown) {
-      logger.error(`Failed setViewportSize: ${url}\n${error}`, {
-        error: error as any
-      })
-      return
-    }
 
-    try {
-      await page.goto(url, {
-        timeout: 8000
-      })
-      await page.waitForSelector('.transparent-bg', { timeout: 2000 })
-    } catch (error) {
-      if (error instanceof Error && error.name === 'TimeoutError') {
-        logger.error(`Timeout while loading page: ${url}`)
+      await page.goto(url)
+
+      const hasImages = view.nodes.some(n => isString(n.icon))
+      if (hasImages) {
+        await waitAllImages(page)
       } else {
-        logger.error(`Page loading failed: ${url}\n${error}`, { error: error as any })
+        await page.waitForSelector('.transparent-bg')
       }
-      return
-    }
 
-    try {
-      // Wait for network to be idle (if there images to be loaded)
-      await page.waitForLoadState('networkidle', { timeout: 15000 })
-    } catch (error: unknown) {
-      logger.error(`Timeout while waiting for page load state: ${url}\n${error}`, {
-        error: error as any
-      })
-      return
-    }
-
-    const path = resolve(outputDir, view.relativePath ?? '.', `${view.id}.png`)
-
-    try {
+      const path = resolve(output, view.relativePath ?? '.', `${view.id}.png`)
       await page.screenshot({
         path,
         animations: 'disabled',
-        timeout: 15000,
         omitBackground: true
       })
-    } catch (error: unknown) {
-      logger.error(`Error when taking screenshot: ${url}\n${error}`, { error: error as any })
-      return
+
+      succeed.push({ view, path })
+    } catch (error) {
+      logger.error(`failed on ${view.id}\n${error}`, { error: error as any })
+      // Force to create new page
+      page = undefined
     }
   }
+
+  return succeed
+}
+
+/**
+ * https://stackoverflow.com/questions/77287441/how-to-wait-for-full-rendered-image-in-playwright
+ */
+async function waitAllImages(page: Page) {
+  // Trigger loading of all images
+  let locators = await page.locator('//img').all()
+  if (!locators.length) return
+  // Set up listeners concurrently
+  const promises = locators.map(locator =>
+    locator.evaluate<unknown, HTMLImageElement>(
+      image =>
+        image.complete || new Promise((res, rej) => {
+          image.onload = res
+          image.onerror = rej
+        }),
+      { timeout: 30_000 } // wait max 30s to load image
+    )
+  )
+  // Wait for all once
+  await Promise.allSettled(promises)
 }